Unify to one test
